# WorkspaceViewModel fast node and connector lookup