--- a/src/DynamoCoreWpf/ViewModels/Core/NodeViewModel.cs
+++ b/src/DynamoCoreWpf/ViewModels/Core/NodeViewModel.cs
@@ -1970,7 +1970,7 @@ namespace Dynamo.ViewModels
 
             foreach (var inode in nodes)
             {
-                var current = this.WorkspaceViewModel.Nodes.FirstOrDefault(x => x.NodeLogic == inode);
+                var current = WorkspaceViewModel.FindNode(inode.GUID);
                 if (current != null)
                 {
                     current.RaisePropertyChanged("IsFrozen");